Attempt at actual paddle collision physics
This adds a commented out function to calculate a new angle for the ball
based on where it collides with the paddle. The general idea is that the
ball's relative position to the paddle determines the outgoing angle -
straight out when it hits the middle, increasing up to 60 degrees
towards the edges.

The math here is (mostly) correct, but because we're on a fixed timestep
the physics system gets confused; ocassionally, the ball ends up
"inside" the paddle, and gets ejected at a strange (and often
game-ending) angle. So it needs to be interpolated, or otherwise pushed
away from the paddle reliably.