Fix paddle/wall collision
Player-controlled paddle collision was broken because player input
dictates paddle velocity, not paddle position. Velocity is (generally)
a continuous function over time; since velocity is being determined via
player input (which isn't continuous), it's difficult to make collision
guesses using velocity. Or position (a derivative of velocity).

So this switches player input to control position directly, giving
player-controlled paddles a velocity of (0, 0). There's a caveat: we
rely on paddle velocity to determine a percentage of ball speed after
a paddle/ball collision. So that has been updated to use "instant
velocity", or a comparison of this tick position vs last tick position.

Whatever, it works.